Testing trigonometric polynomials with a pendulum
There is some floating point error in evaluating the equality constraints.

It looks like finding the maximum height of the pendulum works, which is 1- (max -c).

When Xsupp = [], v(x)-gamma contacts the peak-value curve as desired. If Xsupp = [c^2 + s^2 == 1], then dual_rec has additional entries and the v(x)-gamma is inaccurate. Why? What is the structure of v(x)-gamma? And how to read off the values 'beta' in multiple objectives?

Next step: implement this for the dual formulation (Yalmip). Then experiment with swing-up control.